requirements.txt
pyquery==1.4.0
requests==2.21.0
Pillow==8.0.1
